requirements.txt
# Python 3.11.7

torch
torcheval
transformers
tqdm

numpy
pandas
scikit-learn
catboost
scipy
matplotlib
seaborn
bayesian-optimization
umap-learn
umap-learn[plot]
dask-expr

pyYAML
gdown
colorama

nltk
argostranslate
spacy
tiktoken
openai
lingua-language-detector
rank_bm25
